+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 * **\[annotations\]** Made sure that the return value from an `@ObservableRef` annotated method could accept a
   type parameter. The bug was introduced in v0.23 when `Observable` started to take a type parameter but the annotation
   processing code had not been updated to reflect change.
+* **\[core\]** Fix the generics on the `ArezContext.createObservable(...)` methods to pass the type parameter to
+  observable and make sure the type parameter is documented.
 
 ##### Added
 * **\[core\]** Add the `ArezContext.createObservable()` method that will synthesize the observable name if